create custom image file
