+++ lib/core/functions/navigation.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: use_build_context_synchronously
 
 import 'package:flutter/material.dart';
+
 abstract class NavigatorManager {
 //? navigation without delay
   void navigation({
@@ -9,6 +10,7 @@ abstract class NavigatorManager {
   }) {
     Navigator.pushNamed(context, route);
   }
+
 //? navigation with delay time
   void navigationdelay({
     required BuildContext context,
